+++ BIOL_2200_Calc.py
@@ -3,9 +3,9 @@
 # 20% Lab
 # 10% Mastering Biology
 
-exam_avg = int(input("Please enter your exam average: "))
-lab_avg = int(input("Please enter your lab average: "))
-mast_avg =  int(input("Please enter your Mastering Biology average: "))
+exam_avg = float(input("Please enter your exam average: "))
+lab_avg = float(input("Please enter your lab average: "))
+mast_avg =  float(input("Please enter your Mastering Biology average: "))
 
 def final_grade(exam, lab, mast) :
     return (exam*.7)+(lab*.2)+(mast*.1)
